# backend/swagger.yaml
basePath: /
consumes:
- application/json
...
info:
  description: Documentation of Workers API
  title: Workers API
  version: 1.0.0
paths:
...
  /auth/{level}:
    post:
      description: Returns list of administratorzy
      operationId: CreateSession
      parameters:
      - description: 'The wanted access level: 1-Pracownik 2-Administrator 3-AdministratorDB'
        format: int8
        in: path
        name: level
        required: true
        type: integer
        x-go-name: Id
      responses:
        "201":
          $ref: '#/responses/success'
        "400":
          $ref: '#/responses/error400'
        "500":
          $ref: '#/responses/error500'
      tags:
      - Auth
produces:
- application/json
responses:
...
  error400:
    description: Registered error coming from bad client request
  error500:
    description: Registered error coming from internal server error
  success:
    description: Session creation success
    headers:
      SessionID:
        description: |-
          in: cookie
          name: session-id
        type: string
schemes:
- http
swagger: "2.0"